removed elFinder installation question, separated all elFinder functionality to an add-on
