+++ src/BackpackServiceProvider.php
@@ -70,21 +70,6 @@ class BackpackServiceProvider extends ServiceProvider
 
         // register the artisan commands
         $this->commands($this->commands);
-
-        // register the services that are only used for development
-        // if ($this->app->environment() == 'local') {
-        //     if (class_exists('Laracasts\Generators\GeneratorsServiceProvider')) {
-        //         $this->app->register('Laracasts\Generators\GeneratorsServiceProvider');
-        //     }
-        //     if (class_exists('Backpack\Generators\GeneratorsServiceProvider')) {
-        //         $this->app->register('Backpack\Generators\GeneratorsServiceProvider');
-        //     }
-        // }
-
-        // map the elfinder prefix
-        if (! \Config::get('elfinder.route.prefix')) {
-            \Config::set('elfinder.route.prefix', \Config::get('backpack.base.route_prefix').'/elfinder');
-        }
     }
 
     public function registerMiddlewareGroup(Router $router)
@@ -110,12 +95,8 @@ class BackpackServiceProvider extends ServiceProvider
         $backpack_public_assets = [__DIR__.'/public' => public_path()];
         $backpack_lang_files = [__DIR__.'/resources/lang' => resource_path('lang/vendor/backpack')];
         $backpack_config_files = [__DIR__.'/config' => config_path()];
-        $elfinder_files = [
-            __DIR__.'/config/elfinder.php'      => config_path('elfinder.php'),
-            __DIR__.'/resources/views-elfinder' => resource_path('views/vendor/elfinder'),
-        ];
 
-        // sidebar_content view, which is the only view most people need to overwrite
+        // sidebar content views, which are the only views most people need to overwrite
         $backpack_menu_contents_view = [
             __DIR__.'/resources/views/base/inc/sidebar_content.blade.php'      => resource_path('views/vendor/backpack/base/inc/sidebar_content.blade.php'),
             __DIR__.'/resources/views/base/inc/topbar_left_content.blade.php'  => resource_path('views/vendor/backpack/base/inc/topbar_left_content.blade.php'),
@@ -131,7 +112,6 @@ class BackpackServiceProvider extends ServiceProvider
         $minimum = array_merge(
             // $backpack_views,
             // $backpack_lang_files,
-            // $elfinder_files,
             $error_views,
             $backpack_public_assets,
             $backpack_config_files,
@@ -149,7 +129,6 @@ class BackpackServiceProvider extends ServiceProvider
         $this->publishes($backpack_public_assets, 'public');
         $this->publishes($backpack_custom_routes_file, 'custom_routes');
         $this->publishes($gravatar_assets, 'gravatar');
-        $this->publishes($elfinder_files, 'elfinder');
         $this->publishes($minimum, 'minimum');
     }
